adding dependency and configuration for serialization of yaml types through header params, adding clas utils MediaType for a cleaner controller
