+++ src/main/java/br/Startup.java
@@ -1,4 +1,4 @@
-package com.br;
+package br;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
