+++ src/main/java/br/controllers/PersonController.java
@@ -1,9 +1,8 @@
-package com.br.controllers;
+package br.controllers;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,9 +13,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.br.data.vo.v1.PersonVO;
-import com.br.data.vo.v2.PersonVOV2;
-import com.br.services.PersonServices;
+import br.data.vo.v1.PersonVO;
+import br.data.vo.v2.PersonVOV2;
+import br.services.PersonServices;
+import br.util.MediaType;
 
 @RestController
 @RequestMapping("api/person/v1")
@@ -26,35 +26,36 @@ public class PersonController {
 	private PersonServices service;
 
 	@GetMapping(
-		produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
+		produces = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"})
 	public List<PersonVO> findAll() {
 		return service.findAll();
+		
 	}
 	
 	@GetMapping(value = "/{id}",
-		produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
+		produces = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"})
 	public PersonVO findById(@PathVariable(value = "id") Long id){
 		
 		return service.findById(id);
 	}
 
 	@PostMapping(value = "/v2",
-		produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE},
-		consumes = MediaType.APPLICATION_JSON_VALUE)
+		produces = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"},
+		consumes =  {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"})
 	public PersonVOV2 createV2(@RequestBody PersonVOV2 person) {
 		return service.createV2(person);
 	}
 
 	@PostMapping(
-		produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE},
-		consumes = MediaType.APPLICATION_JSON_VALUE)
+		produces = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"},
+		consumes = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"})
 	public PersonVO create(@RequestBody PersonVO person) {
 		return service.create(person);
 	}
 	
 	@PutMapping(
-		produces =  {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE},
-		consumes =  {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
+		produces = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"},
+		consumes = {MediaType.APPLICATION_JSON , MediaType.APPLICATION_XML , "application/x-ymal"})
 	public PersonVO update(@RequestBody PersonVO person) {
 		return service.update(person);
 	}
