+++ src/main/java/br/exceptions/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.br.exceptions;
+package br.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
