+++ src/main/java/br/model/Person.java
@@ -1,26 +1,38 @@
-package com.br.data.vo.v2;
+package br.model;
 
 import java.io.Serializable;
-import java.util.Date;
 import java.util.Objects;
 
-public class PersonVOV2 implements Serializable{
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name= "person")
+public class Person implements Serializable{
 
 	private static final long serialVersionUID = 1L;
 	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
+	@Column(name = "first_name", nullable = false, length = 80)
 	private String  firstName;
 	
+	@Column(name = "last_name", nullable = false, length = 80)
 	private String  lastName;
 	
+	@Column(nullable = false, length = 100)
 	private String  address;
 	
+	@Column(nullable = false, length = 6)
 	private String  gender;
 	
-	private Date birthdDay;
-	
-	public PersonVOV2() {}
+	public Person() {}
 
 	public String getFirstName() {
 		return firstName;
@@ -62,17 +74,9 @@ public class PersonVOV2 implements Serializable{
 		this.id = id;
 	}
 
-	public Date getBirthdDay() {
-		return birthdDay;
-	}
-
-	public void setBirthdDay(Date birthdDay) {
-		this.birthdDay = birthdDay;
-	}
-
 	@Override
 	public int hashCode() {
-		return Objects.hash(address, birthdDay, firstName, gender, id, lastName);
+		return Objects.hash(address, firstName, gender, id, lastName);
 	}
 
 	@Override
@@ -83,13 +87,11 @@ public class PersonVOV2 implements Serializable{
 			return false;
 		if (getClass() != obj.getClass())
 			return false;
-		PersonVOV2 other = (PersonVOV2) obj;
-		return Objects.equals(address, other.address) && Objects.equals(birthdDay, other.birthdDay)
-				&& Objects.equals(firstName, other.firstName) && Objects.equals(gender, other.gender)
-				&& Objects.equals(id, other.id) && Objects.equals(lastName, other.lastName);
+		Person other = (Person) obj;
+		return Objects.equals(address, other.address) && Objects.equals(firstName, other.firstName)
+				&& Objects.equals(gender, other.gender) && Objects.equals(id, other.id)
+				&& Objects.equals(lastName, other.lastName);
 	}
-
-	
 	
 	
 }
