+++ src/main/java/br/serializatoin/converter/YamlJackson2HttpMessageConverter.java
@@ -1,0 +1,16 @@
+package br.serializatoin.converter;
+
+import org.springframework.http.MediaType;
+import org.springframework.http.converter.json.AbstractJackson2HttpMessageConverter;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
+
+public class YamlJackson2HttpMessageConverter extends AbstractJackson2HttpMessageConverter {
+
+	protected YamlJackson2HttpMessageConverter() {
+		super(new YAMLMapper().setSerializationInclusion(JsonInclude.Include.NON_NULL),
+				MediaType.parseMediaType("application/x-yaml"));
+	}
+
+}
