+++ src/main/java/br/util/MediaType.java
@@ -1,0 +1,8 @@
+package br.util;
+
+public class MediaType {
+	public static final String APPLICATION_JSON ="application/json";
+	public static final String APPLICATION_XML ="application/xml";
+	public static final String APPLICATION_YML ="application/x-yaml";
+
+}
